refactor: reorganize backend from layer-grouped to feature-grouped
Per ARCHITECTURE.md: "Code is organized by feature modules rather than
technical layers."

Structure changes:
- app/api/ + app/models/ (layer-grouped) → app/auth/, app/todos/,
  app/ai/, app/health/ (feature-grouped)
- Each feature owns its models, schemas, and routes
- auth/ merges sessions + users routes into a single routes.py
- Token schema de-duplicated into auth/schemas.py

RESTful route improvements:
- GET /ai/suggestions/{todo_id} → GET /todos/{todo_id}/suggestions
- GET /ai/summarize/{list_id} → GET /todos/lists/{list_id}/summary
- Removed stub-only AI endpoints from ai/ (now sub-resources of todos)
