+++ backend/app/auth/schemas.py
@@ -1,0 +1,36 @@
+"""Auth-related Pydantic schemas."""
+
+from datetime import datetime
+
+from pydantic import BaseModel, field_validator
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class LoginRequest(BaseModel):
+    email: str
+    password: str
+
+
+class UserCreate(BaseModel):
+    email: str
+    password: str
+
+    @field_validator("email")
+    @classmethod
+    def email_not_empty(cls, v: str) -> str:
+        if not v.strip():
+            raise ValueError("Email must not be empty")
+        return v.strip()
+
+
+class UserRead(BaseModel):
+    id: str
+    email: str
+    is_active: bool
+    created_at: datetime
+
+    model_config = {"from_attributes": True}
