--- backend/app/models/__init__.py
@@ -1,4 +1,0 @@
-from app.models.todo import Location, RecurrenceRule, Reminder, Todo, TodoList
-from app.models.user import User
-
-__all__ = ["Todo", "TodoList", "Reminder", "Location", "RecurrenceRule", "User"]
